move teradata tasks to separate files
