+++ dumper/app/src/main/java/com/google/edwmigration/dumper/application/dumper/connector/teradata/TeradataAssessmentLogsJdbcTask.java
@@ -1,0 +1,144 @@
+package com.google.edwmigration.dumper.application.dumper.connector.teradata;
+
+import com.google.edwmigration.dumper.application.dumper.connector.ZonedInterval;
+import com.google.edwmigration.dumper.application.dumper.connector.teradata.TeradataLogsConnector.SharedState;
+import java.util.List;
+import java.util.function.Predicate;
+import javax.annotation.Nonnull;
+
+public class TeradataAssessmentLogsJdbcTask extends TeradataLogsJdbcTask {
+        static final String[] EXPRESSIONS = new String[]{
+                "L.QueryID",
+                "ST.SQLRowNo",
+                "ST.SQLTextInfo",
+                "L.AbortFlag",
+                "L.AcctString",
+                "L.AcctStringDate",
+                "L.AcctStringHour",
+                "L.AcctStringTime",
+                "L.AMPCPUTime",
+                "L.AMPCPUTimeNorm",
+                "L.AppID",
+                "L.CacheFlag",
+                "L.CalendarName",
+                "L.CallNestingLevel",
+                "L.CheckpointNum",
+                "L.ClientAddr",
+                "L.ClientID",
+                "L.CollectTimeStamp AT TIME ZONE INTERVAL '0:00' HOUR TO MINUTE AS \"CollectTimeStamp\"",
+                "L.CPUDecayLevel",
+                "L.DataCollectAlg",
+                "L.DBQLStatus",
+                "L.DefaultDatabase",
+                "L.DelayTime",
+                "L.DisCPUTime",
+                "L.DisCPUTimeNorm",
+                "L.ErrorCode",
+                "L.EstMaxRowCount",
+                "L.EstMaxStepTime",
+                "L.EstProcTime",
+                "L.EstResultRows",
+                "L.ExpandAcctString",
+                "L.FirstRespTime AT TIME ZONE INTERVAL '0:00' HOUR TO MINUTE AS \"FirstRespTime\"",
+                "L.FirstStepTime AT TIME ZONE INTERVAL '0:00' HOUR TO MINUTE AS \"FirstStepTime\"",
+                "L.FlexThrottle",
+                "L.ImpactSpool",
+                "L.InternalRequestNum",
+                "L.IODecayLevel",
+                "L.IterationCount",
+                "L.KeepFlag",
+                "L.LastRespTime",
+                "L.LockDelay",
+                "L.LockLevel",
+                "L.LogicalHostID",
+                "L.LogonDateTime AT TIME ZONE INTERVAL '0:00' HOUR TO MINUTE AS \"LogonDateTime\"",
+                "L.LogonSource",
+                "L.LSN",
+                "L.MaxAMPCPUTime",
+                "L.MaxAMPCPUTimeNorm",
+                "L.MaxAmpIO",
+                "L.MaxCPUAmpNumber",
+                "L.MaxCPUAmpNumberNorm",
+                "L.MaxIOAmpNumber",
+                "L.MaxNumMapAMPs",
+                "L.MaxOneMBRowSize",
+                "L.MaxStepMemory",
+                "L.MaxStepsInPar",
+                "L.MinAmpCPUTime",
+                "L.MinAmpCPUTimeNorm",
+                "L.MinAmpIO",
+                "L.MinNumMapAMPs",
+                "L.MinRespHoldTime",
+                "L.NumFragments",
+                "L.NumOfActiveAMPs",
+                "L.NumRequestCtx",
+                "L.NumResultOneMBRows",
+                "L.NumResultRows",
+                "L.NumSteps",
+                "L.NumStepswPar",
+                "L.ParamQuery",
+                "L.ParserCPUTime",
+                "L.ParserCPUTimeNorm",
+                "L.ParserExpReq",
+                "L.PersistentSpool",
+                "L.ProcID",
+                "L.ProfileID",
+                "L.ProfileName",
+                "L.ProxyRole",
+                "L.ProxyUser",
+                "L.ProxyUserID",
+                "L.QueryBand",
+                "L.QueryRedriven",
+                "L.QueryText",
+                "L.ReDriveKind",
+                "L.RemoteQuery",
+                "L.ReqIOKB",
+                "L.ReqPhysIO",
+                "L.ReqPhysIOKB",
+                "L.RequestMode",
+                "L.RequestNum",
+                "L.SeqRespTime",
+                "L.SessionID",
+                "L.SessionTemporalQualifier",
+                "L.SpoolUsage",
+                "L.StartTime AT TIME ZONE INTERVAL '0:00' HOUR TO MINUTE AS \"StartTime\"",
+                "L.StatementGroup",
+                "L.Statements",
+                "L.StatementType",
+                "L.SysDefNumMapAMPs",
+                "L.TacticalCPUException",
+                "L.TacticalIOException",
+                "L.TDWMEstMemUsage",
+                "L.ThrottleBypassed",
+                "L.TotalFirstRespTime",
+                "L.TotalIOCount",
+                "L.TotalServerByteCount",
+                "L.TTGranularity",
+                "L.TxnMode",
+                "L.TxnUniq",
+                "L.UnitySQL",
+                "L.UnityTime",
+                "L.UsedIota",
+                "L.UserID",
+                "L.UserName",
+                "L.UtilityByteCount",
+                "L.UtilityInfoAvailable",
+                "L.UtilityRowCount",
+                "L.VHLogicalIO",
+                "L.VHLogicalIOKB",
+                "L.VHPhysIO",
+                "L.VHPhysIOKB",
+                "L.WarningOnly",
+                "L.WDName"
+        };
+
+        public TeradataAssessmentLogsJdbcTask(@Nonnull String targetPath, SharedState state, String logTable, String queryTable, List<String> conditions, ZonedInterval interval, List<String> orderBy) {
+            super(targetPath, state, logTable, queryTable, conditions, interval, orderBy);
+        }
+
+        @Nonnull
+        @Override
+        String getSql(@Nonnull Predicate<? super String> predicate) {
+            return getSql(predicate, EXPRESSIONS);
+        }
+    }
